changed modals

+++ checkin/models.py
@@ -16,7 +16,7 @@ class Visitors(models.Model):
 class Checkins(models.Model):
     visitor = models.ForeignKey(Visitors, on_delete=models.CASCADE)
     #use floatfield
-    temperature = models.IntegerField()
+    temperature = models.FloatField()
     created_date = models.DateTimeField(default=timezone.now)
     def checkin(self):
         self.save()
